添加了阅读次数的显示和shift+C复制编辑笔记代码的功能

+++ run.sh
@@ -4,7 +4,17 @@ appName="blog-java"
 app=$(pgrep -f $appName)
 if test ${app}null != "null"
 then
-    kill -9 $app
+    # 调用应用接口执行保存操作
+    status=$(curl http://localhost:8080/writeBlogVisitTimes)
+    if test ${status} == 'success'
+    then
+        # 如果保存成功，则停止应用
+        kill -9 $app
+    else
+        # 如果保存失败，停止脚本
+        echo 'blog访问次数保存失败'
+        exit
+    fi
 fi
 
 # 启动应用
